refactoring and clean-up
