+++ src/components/Landing/Landing.js
@@ -1,0 +1,4 @@
+import React, {Component} from 'react';
+import { View, Text, StyleSheet, TouchableOpacity } from 'react-native';
+
+import LandingStyles from './LandingStyles';
