+++ src/components/Landing/LandingStyles.js
@@ -1,0 +1,7 @@
+import {StyleSheet} from 'react-native';
+
+export default StyleSheet.create({
+    container: {
+        flex: 1
+    }
+})
