+++ src/components/Scanner/ScannerStyles.js
